[#99] Split `Toml.Parser` file into smaller files
Refactored the Parser module:

- Toml.Parser.Core:
    Exposes the fundamental parsing functions. This allows easily switching between
    different parser libraries for testing and benchmarking. The current implementation is
    based on MegaParsec.

- Toml.Parser.Value:
    Parsers for the basic values of the language.

- Toml.Parser.TOML:
    Parser for the structures of the language.

- Toml.Parser:
    Exposed parser API: exceptions and high level functions.

Updated the unit tests for the parser module. No new tests were implemented.

Updated the cabal file to include the new modules.
